Update sanity calculation methods for accuracy and precision
Replaced BlockPos instantiation with a more precise factory method and adjusted sanity return value scaling in commands. These changes improve positioning accuracy and ensure sanity values are consistently represented as percentages.
Also fixed a bug with `/sanity get <target>` returning a float between -1 and 1 instead of an int between -100 and 100 to the user in-game.